move fix to subclass
